+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Bug Fixes
 
+- [boltdb] Properly handle blank keys in iterators
+
 - [goleveldb] [\#58](https://github.com/tendermint/tm-db/pull/58) Make `Batch.Close()` actually remove the batch contents
 
 ## 0.4.1